SBT build: bumping to scala 2.8.2
This fixes an issue we've been having when two parallel threads talks
with the ClusterNotificationManager. For example, a JVM with both client
and server would have the ClusterClient call
ClusterNotificationMessages.GetCurrentNodes, but every now and then one
of the calls failed with "Received unknown message: GetCurrentNodes",
leaving the ClusterClient blocking forever.
Most likely this was due to https://issues.scala-lang.org/browse/SI-1591
which is fixed in 2.8.2

Headsup: with this "bugfix" version of scala some boxing changes seems
to have happened.. If you use scala libraries in java code, for example
Option[Long] is no longer java Option<Long> but now Option<Object>..??