Initial commit of Dark Canary channel handler.
The basic strategy here is to insert a downstream channel handler before the default channel handler. This
new handler will inspect the request to see if it is going to a mirrored host. If it is, it copies the request
with, modifies the target host, stores the request id, and then sends the new request to the io client.

We also install an upstream handler. This handler gets responses beofre the default channel handler. The new
handler checks to see if a response is for a mirrored request. If it is, it stops propagation of the response,
hence ensuring that the dark responses can never reach the production client.