Added support for building against Prototbuf 2.6
Since ProtoUtils does some speedup-magic, special care had to be taken.
The ByteString(byte[]) constructor is gone in 2.6. Instead a new
LiteralByteString class exists.
ProtoUtils will auto-detect if this exists, and use that approach
instead.

Note that the builder must manually rebuild the protocols in
cluster/src/main/protobuf, since the generated files are NOT compatible!